proje/requirements.txt
flask
opencv-python
torch
numpy
paho-mqtt
pandas
requests
Pillow
seaborn
matplotlib
scipy
pyyaml
tqdm
